+++ test/time_zone_info/naive_datetime_util_test.exs
@@ -1,3 +1,0 @@
-defmodule TimeZoneInfo.NaiveDateTimeUtilTest do
-  use ExUnit.Case, async: true
-end