+++ lib/time_zone_info/transformer/rule.ex
@@ -3,9 +3,8 @@ defmodule TimeZoneInfo.Transformer.Rule do
   This module handles and transforms the IANA rules.
   """
 
-  alias TimeZoneInfo.GregorianSeconds
+  alias TimeZoneInfo.IanaDateTime
   alias TimeZoneInfo.IanaParser
-  alias TimeZoneInfo.NaiveDateTimeUtil, as: NaiveDateTime
   alias TimeZoneInfo.Transformer.RuleSet
 
   @doc """
@@ -38,15 +37,12 @@ defmodule TimeZoneInfo.Transformer.Rule do
           end
 
         Enum.into(from..to, [], fn year ->
-          at =
-            year
-            |> NaiveDateTime.from_iana(rule[:in], rule[:on], rule[:at])
-            |> GregorianSeconds.from_naive()
+          at = IanaDateTime.to_gregorian_seconds(year, rule[:in], rule[:on], rule[:at])
 
           {at, {rule[:time_standard], rule[:std_offset], rule[:letters]}}
         end)
       end)
-      |> GregorianSeconds.sort()
+      |> Enum.sort_by(&elem(&1, 0))
 
     {_, first} = first_standard(rule_set)
     [{-1, first} | rule_set]