+++ test/time_zone_info/data_store/erlang_term_storage_test.exs
@@ -46,18 +46,22 @@ defmodule TimeZoneInfo.DataStore.ErlangTermStorageTest do
   end
 
   test "info" do
-    assert DataStore.info() == %{
-             version: "2019c",
-             time_zones: 387,
-             links: 86,
-             tables: %{
-               time_zone_info: [size: 1, memory: 317],
-               time_zone_info_links: [size: 86, memory: 1625],
-               time_zone_info_rules: [size: 29, memory: 2160],
-               time_zone_info_time_zones: [size: 3, memory: 6274],
-               time_zone_info_transitions: [size: 387, memory: 323_687]
-             }
-           }
+    assert %{version: version, time_zones: time_zones, links: links, tables: tables} =
+             DataStore.info()
+
+    assert version == "2019c"
+    assert time_zones == 387
+    assert links == 86
+    assert get_in(tables, [:time_zone_info, :size]) == 1
+    assert get_in(tables, [:time_zone_info, :memory]) > 300
+    assert get_in(tables, [:time_zone_info_links, :size]) == 86
+    assert get_in(tables, [:time_zone_info_links, :memory]) > 1600
+    assert get_in(tables, [:time_zone_info_rules, :size]) == 29
+    assert get_in(tables, [:time_zone_info_rules, :memory]) > 2100
+    assert get_in(tables, [:time_zone_info_time_zones, :size]) == 3
+    assert get_in(tables, [:time_zone_info_time_zones, :memory]) > 6200
+    assert get_in(tables, [:time_zone_info_transitions, :size]) == 387
+    assert get_in(tables, [:time_zone_info_transitions, :memory]) > 323_000
   end
 
   describe "time_zone_period_from_utc_iso_days/2" do