+++ test/time_zone_info/data_store/persistent_term_test.exs
@@ -6,7 +6,7 @@ defmodule TimeZoneInfo.DataStore.PersistentTermTest do
     alias TimeZoneInfo.DataStore
 
     test "get_transitions/1" do
-      assert {:ok, zone_states} = TimeZoneInfo.DataStore.get_transitions("Pacific/Auckland")
+      assert {:ok, zone_states} = DataStore.get_transitions("Pacific/Auckland")
 
       assert Enum.take(zone_states, 3) == [
                {64_241_906_400, {43200, "NZ", {:template, "NZ%sT"}}},
@@ -16,7 +16,7 @@ defmodule TimeZoneInfo.DataStore.PersistentTermTest do
     end
 
     test "get_rules/1" do
-      assert {:ok, rules} = TimeZoneInfo.DataStore.get_rules("NZ")
+      assert {:ok, rules} = DataStore.get_rules("NZ")
 
       assert rules == [
                {{4, [day: 1, op: :ge, day_of_week: 7], {2, 0, 0}}, :standard, 0, "S"},