+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Adding HTTP status code to the downloader.
 - Adding optional `info/0` callback to the `DataStore` behaviour.
 - Updating default `lookahead` to 15 years.
-
+- Refactoring date time handling.
 
 ## 0.2.0 - 2020/04/08